fix up int literal checker bug

+++ src/edu/mit/compilers/ir/expression/literal/IRIntLiteral.java
@@ -3,7 +3,7 @@ package edu.mit.compilers.ir.expression.literal;
 import antlr.Token;
 import edu.mit.compilers.ir.common.IRVisitor;
 
-public class IRIntLiteral extends IRLiteral<Integer> {
+public class IRIntLiteral extends IRLiteral<Long> {
 	
 	public IRIntLiteral(Token token) {
 		super("IRIntLiteral", token);
